Payments include whether they were Reconciled
The Payments XML returned from the invoice does not include reconciliation
information, but the XML from the /Payments endpoint does.

This treats it as a pretty optional paramater, in that if it's falsey it
doesn't even pass it on. This may not be desired behavior (for instance, if
someone want to 'un-reconciled' a payment); however I didn't want UpdatePayment
to start marking payments as unreconciled just because reconciled was nil

We may want to do a more explicit comparison?