Payment#reconciled is set only if IsReconciled XML is 'true'
We had seen cases where `IsReconciled` did not come in for Payments, so we
assumed that any time IsReconciled *is* set then the payment was reconciled.

This is not, apparently, true. IsReconciled can sometimes be a string value of
`"false"`; so we are doing a string comparison to ensure that the IsReconciled
is the string value of `"true"`