Link to accurate instructions to generate certificate
The instructions to generate a keypair / certificate did not work for me and I kept receiving invalid signature errors.  After some time shooting in the dark I tried the instructions on the xero docs site and everything worked.  I'm not sure why but there are slight differences in these instructions and the github local ones seem not to work on my environment (mac 10.11.6, ruby 2.4).